Fully Functioning
There are still a few bugs in places that need to be fixed, but for the
most part the app is complete and functioning as it should.

Final report draft has been updated to include front-end back-end
discussion, screenshots, and a few other small updates like editing the
order of some attributes in the DDLs to match the database.